StyleCop対応（仮対応）
