+++ CSharpSample/CSharpSample.csproj
@@ -74,6 +74,11 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Common\Types\GenericsInterfaceSample01.cs" />
+    <Compile Include="Common\Types\GenericsInterfaceSample02.cs" />
+    <Compile Include="Common\Types\Person.cs" />
+    <Compile Include="Common\Types\TypeIN.cs" />
+    <Compile Include="Common\Types\TypeOut.cs" />
     <Compile Include="DataFactory\BaseSampleData.cs" />
     <Compile Include="DataFactory\DataFactory.cs" />
     <Compile Include="DataFactory\SampleData001.cs" />
@@ -106,7 +111,9 @@
     <None Include="App.config" />
     <None Include="packages.config" />
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Folder Include="Common\Interfaces\" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
