+++ CSharpSample/Common/Types/TypeIN.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CSharpSample.Common.Types
+{
+    /// <summary>
+    /// Genericsサンプルで利用するクラス1(TypeIn:入力されるクラスとして用意)
+    /// </summary>
+    public class TypeIN
+    {
+        /// <summary>
+        /// コンストラクタ
+        /// </summary>
+        public TypeIN()
+        {
+            Console.WriteLine("Constructor : TypeIN");
+        }
+    }
+}
