+++ CSharpSample/Common/Types/TypeOut.cs
@@ -1,0 +1,18 @@
+﻿using System;
+
+namespace CSharpSample.Common.Types
+{
+    /// <summary>
+    /// Genericsサンプルで利用するクラス1(TypeOut:出力されるくらすとして用意)
+    /// </summary>
+    public class TypeOut
+    {
+        /// <summary>
+        /// コンストラクタ
+        /// </summary>
+        public TypeOut()
+        {
+            Console.WriteLine("Constructor : TypeOut");
+        }
+    }
+}
